requirements.txt
flask==3.0.0
openai-whisper==20231117
torch>=2.0.0
torchaudio>=2.0.0
soundfile>=0.12.1
python-multipart>=0.0.6
